+++ cells/std/cli/keys/keys.go
@@ -1,4 +1,4 @@
-package main
+package keys
 
 import (
 	"github.com/charmbracelet/bubbles/key"
