+++ cells/std/cli/styles/styles.go
@@ -1,0 +1,28 @@
+package styles
+
+import (
+	"github.com/charmbracelet/lipgloss"
+)
+
+var (
+	Highlight = lipgloss.AdaptiveColor{Light: "#874BFD", Dark: "#7D56F4"}
+	AppStyle  = lipgloss.NewStyle().Padding(1, 2)
+
+	TargetStyle = lipgloss.NewStyle().
+			BorderStyle(lipgloss.NormalBorder()).
+			BorderForeground(Highlight)
+
+	ActionStyle = lipgloss.NewStyle().
+			BorderStyle(lipgloss.NormalBorder()).
+			BorderForeground(Highlight)
+
+	ReadmeStyle = lipgloss.NewStyle().
+		// BorderStyle(lipgloss.NormalBorder()).
+		BorderForeground(Highlight)
+
+	LegendStyle = lipgloss.NewStyle().Padding(1, 0, 0, 2)
+
+	TitleStyle = lipgloss.NewStyle().
+			Foreground(Highlight).Bold(true).
+			Padding(1, 1)
+)
